chore: reorganize project folders for better structure
