+++ controllers/productControllers/getAllProductsControllers.js
@@ -1,0 +1,8 @@
+const productsService = require('../../services/productsService');
+
+const getAll = async (_req, res) => {
+  const products = await productsService.getAll();
+  res.status(200).json(products);
+};
+
+module.exports = getAll;
