+++ package.json
@@ -4,7 +4,7 @@
   "description": "a CRUD to sales and products",
   "main": "index.js",
   "scripts": {
-    "test": "jest --runInBand",
+    "test": "jest --runInBand --watchAll",
     "test:mocha": "nyc --all --include models --include services --include controllers mocha test/unit/*.js --exit",
     "start": "node index.js",
     "debug": "nodemon index.js",
@@ -28,7 +28,7 @@
     "express": "^4.17.1",
     "express-rescue": "^1.1.26",
     "frisby": "^2.1.2",
-    "jest": "^26.4.1",
+    "jest": "^26.6.3",
     "joi": "^17.9.2",
     "mysql2": "^2.3.0"
   },
