+++ routes/productsRoutes.js
@@ -1,13 +1,19 @@
 const express = require('express');
 const rescue = require('express-rescue');
-const productsController = require('../controllers/productsController');
+
+const validateProducts = require('../controllers/productControllers/validadeProductsControllers')
+const createProducts = require('../controllers/productControllers/createProductsControllers')
+const getAllProducts = require('../controllers/productControllers/getAllProductsControllers')
+const getByIdProducts = require('../controllers/productControllers/getByIdProductsControllers')
+const updateProducts = require('../controllers/productControllers/updateProductsControllers')
+const removeProducts = require('../controllers/productControllers/removeProductsControllers')
 
 const router = express.Router();
 
-router.get('/', rescue(productsController.getAll));
-router.get('/:id', rescue(productsController.getById));
-router.put('/:id', rescue(productsController.validateProducts), rescue(productsController.update));
-router.delete('/:id', rescue(productsController.del));
-router.post('/', rescue(productsController.validateProducts), rescue(productsController.create));
+router.get('/', rescue(getAllProducts));
+router.get('/:id', rescue(getByIdProducts));
+router.put('/:id', rescue(validateProducts), rescue(updateProducts));
+router.delete('/:id', rescue(removeProducts));
+router.post('/', rescue(validateProducts), rescue(createProducts));
 
 module.exports = router;
